% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_accordion_multi.R
\name{bs_accordion_multi}
\alias{bs_accordion_multi}
\title{Allow multiple \code{bs_accordion} sections to be open, and specify which ones start as open}
\usage{
bs_accordion_multi(X, multi = TRUE, open = 1)
}
\arguments{
\item{X}{shinyBS::bs_accordion object}

\item{multi}{logical Should more than one panel be permitted to be open at one time.}

\item{open}{integer vector Indices of panels that should be open when the page loads.}
}
\value{
. Modified shinyBS::bs_accordion object
}
\description{
Allow multiple \code{bs_accordion} sections to be open, and specify which ones start as open
}
\examples{

if(interactive())

shinyApp(
  ui = fluidPage(

    tags$header(
      tags$style(
      HTML(
        '
        .panel-heading.collapsed > .panel-title:before {
            float: right !important;
            padding-right: 5px;
            content:"+"
            }

         .panel-heading > .panel-title:before {
           float: right !important;
           font-family: FontAwesome;
           padding-right: 5px;
           content:"-"
         }
         '
        )
      )
    ),

    tags$h2("Multiple Open Panels, Panels 1 and 3 open at page load"),
    bs_accordion(id = "beatles") \%>\%
      bs_set_opts(panel_type = "success", use_heading_link = TRUE) \%>\%
      bs_append(title = "John Lennon", content = "Rhythm guitar, vocals") \%>\%
      bs_append(title = "Paul McCartney", content = "Bass guitar, vocals") \%>\%
      bs_append(title = "George Harrison", content = "Lead guitar, vocals") \%>\%
      bs_append(title = "Ringo Starr", content = "Drums, vocals") \%>\%
      bs_accordion_multi(
        multi=TRUE,
        open=c(1,3)
        ),

   tags$h2("One Open Panel, No panels open on page load."),
    bs_accordion(id = "fruit") \%>\%
      bs_set_opts(panel_type = "info") \%>\%
      bs_append(title = "Apples", content = "An apple a day keeps the doctor away.") \%>\%
      bs_append(title = "Bannana", content = "Watch out for bannana peels!") \%>\%
      bs_append(title = "Kumquat", content = "What is a kumquat?!") \%>\%
      bs_accordion_multi(
        multi=FALSE,
        open=c()
        )
      ),

  server = function(input, output) {}
  )



}
