% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_sidebar.R
\name{bs_accordion_sidebar}
\alias{bs_accordion_sidebar}
\title{framework for accordion-sidebar}
\usage{
bs_accordion_sidebar(id, spec_side = c(width = 4, offset = 0),
  spec_main = c(width = 8, offset = 0), position = c("left", "right"))
}
\arguments{
\item{id}{character, id for div and root it for included elements}

\item{spec_side}{numeric, column specification for sidebar panels}

\item{spec_main}{numeric, column specification for main panels}

\item{position}{character, indicates where to put the sidebar panels with
repspect to the main panels}
}
\value{
\code{bsplus_accsidebar} object
}
\description{
framework for accordion-sidebar
}
\examples{
bs_accordion_sidebar(id = "meet_the_beatles")

}

