% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{bs_attach_collapse}
\alias{bs_attach_collapse}
\title{Attaches a collapsible div's id to a tag}
\usage{
bs_attach_collapse(tag, id_collapse)
}
\arguments{
\item{tag}{htmltools \code{\link[htmltools]{tag}},}

\item{id_collapse}{character, id for the collapsible div}
}
\value{
\code{\link[htmltools]{tag}}, modified copy of \code{tag}
}
\description{
Typical to attach to a button
or link tag (see \code{\link[htmltools]{builder}}).
}
\details{
You may attach a collapsible div's id to more than one tag.
}
\examples{
library("htmltools")
bs_collapse(id = "id_yeah", tags$p("Yeah Yeah Yeah"))
tags$button("She Loves You") \%>\%
  bs_collapse_attach("id_yeah")

}
\seealso{
\url{https://getbootstrap.com/javascript/#collapse}

Other collapse.functions: \code{\link{bs_collapse}}
}

