% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_attr.R
\docType{methods}
\name{bs_attr}
\alias{bs_attr}
\alias{bs_attr.default}
\alias{bs_attr.logical}
\alias{bs_attr.list}
\alias{bs_attr.Duration}
\alias{bs_attr}
\alias{bs_attr,logical-method}
\alias{bs_attr,list-method}
\alias{bs_attr,Duration-method}
\title{Convert a value to be used as an html attribute}
\usage{
bs_attr(x)

bs_attr.default(x)

bs_attr.logical(x)

bs_attr.list(x)

bs_attr.Duration(x)

bs_attr(x)

\S4method{bs_attr}{logical}(x)

\S4method{bs_attr}{list}(x)

\S4method{bs_attr}{Duration}(x)
}
\arguments{
\item{x}{value to be converted}
}
\value{
character
}
\description{
\itemize{
\item character vectors collapsed to a space-delimited character string
\item logicals are converted to "true" or "false"
\item lubridate durations are converted to numeric (milliseconds)
\item lists are converted to JSON.
\strong{Limitation}: \code{\link[lubridate:duration]{lubridate::duration()}} objects within lists
will not translate correctly
}
}
\keyword{internal}
