% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R, R/carousel.R
\name{bs_append.bsplus_carousel}
\alias{bs_append.bsplus_carousel}
\alias{bs_carousel}
\alias{bs_carousel_caption}
\alias{bs_carousel_image}
\title{Create and append to a carousel}
\usage{
\method{bs_append}{bsplus_carousel}(tag, content, caption = NULL, ...)

bs_carousel(id, use_indicators = FALSE, use_controls = TRUE)

bs_carousel_caption(title = NULL, body = NULL)

bs_carousel_image(...)
}
\arguments{
\item{tag}{\code{htmltools::\link[htmltools]{tag}},
carousel div to which to append a panel}

\item{content}{\code{htmltools::\link[htmltools]{tag}},
content for the slide}

\item{caption}{\code{htmltools::\link[htmltools]{tag}},
caption for the slide}

\item{...}{additional arguments passed to \code{htmltools::\link[htmltools]{tag}$img}}

\item{id}{character, unique id for carousel div}

\item{use_indicators}{logical, denotes use of indicators (dots)}

\item{use_controls}{logical, denotes use of controls (chevrons at sides)}

\item{title}{character, caption title}

\item{body}{character, caption body}
}
\value{
\code{htmltools::\link[htmltools]{tag}} div for carousel caption

\code{htmltools::\link[htmltools]{tag}$img}
}
\description{
A carousel is a means to show a group of images, one at a time.
}
\seealso{
\url{http://getbootstrap.com/javascript/#carousel}
}

