% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{bs_carousel_caption}
\alias{bs_carousel_caption}
\title{Carousel caption}
\usage{
bs_carousel_caption(title = NULL, body = NULL)
}
\arguments{
\item{title}{character, caption title}

\item{body}{character, caption body}
}
\value{
\code{htmltools::\link[htmltools]{tag}} \code{<div/>} for carousel caption
}
\description{
Helper function to generate HTML for a carousel caption.
}
\seealso{
\code{\link{bs_carousel}}
}

