% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{bs_carousel_image}
\alias{bs_carousel_image}
\title{Carousel image}
\usage{
bs_carousel_image(...)
}
\arguments{
\item{...}{additional arguments passed to
\code{htmltools::\link[htmltools]{tag}$img}, typically includes \code{src}}
}
\value{
\code{htmltools::\link[htmltools]{tag}}, \verb{</img>}
}
\description{
Helper function to generate HTML for a carousel image.
}
\details{
This function wraps \code{htmltools::\link[htmltools]{tag}$img},
but adding a class to center the image in the carousel.
}
\seealso{
\code{\link{bs_carousel}}
}
