% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{bs_collapse}
\alias{bs_attach_collapse}
\alias{bs_collapse}
\title{Create and attach a collapsible div}
\usage{
bs_collapse(id, content = NULL, show = FALSE)

bs_attach_collapse(tag, id_collapse)
}
\arguments{
\item{id}{character, unique id for modal window to create}

\item{content}{character (HTML) or \code{htmltools::\link[htmltools]{tagList}},
content to be put into the div}

\item{show}{logical, indicates if element is shown initially}

\item{tag}{character, HTML element to which to attach collapsible div}

\item{id_collapse}{character, unique id of collapsible div to attach}
}
\value{
\code{htmltools::\link[htmltools]{tag}} div
}
\description{
When a page is renedered, a collapsible div will be hidden initially
(by default), but can be revealed  using a control to which you will
have attached this div's id.
}
\details{
It is your responsibility to ensure that id is unique
among html elements in your page. If you have non-unique id's, strange things may
happen to your page.

Keep in mind that you can attach a collapsible div to more than one element.
}
\examples{
library("htmltools")

bs_collapse(id = "id_yeah", tags$p("Yeah Yeah Yeah"))

tags$button("She Loves You") \%>\%
  bs_attach_collapse("id_yeah")

}
\seealso{
\url{https://getbootstrap.com/javascript/#collapse}
}

