% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_popover.R
\name{bs_embed_popover}
\alias{bs_embed_popover}
\alias{use_bs_popover}
\title{Embed a popover into an html element}
\usage{
bs_embed_popover(tag, title = NULL, content = NULL, placement = NULL, ...)

use_bs_popover()
}
\arguments{
\item{tag}{shinytag}

\item{title}{character,}

\item{content}{character,}

\item{placement}{character,}

\item{...}{other named arguments, passed to bs_set_data()}
}
\value{
tag
}
\description{
Embed a popover into an html element
}
\examples{
library("htmltools")
tags$button(type = "button", class = "btn btn-default", "I'm a button") \%>\%
  bs_embed_popover(title = "I'm a popover", content = "Really!")
}

