% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_tooltip.R
\name{bs_embed_tooltip}
\alias{bs_embed_tooltip}
\title{Embed a tooltip into an html element}
\usage{
bs_embed_tooltip(tag, title = "", placement = c("right", "auto"), ...)
}
\arguments{
\item{tag}{shinytag}

\item{title}{character,}

\item{placement}{character,}

\item{...}{other named arguments, passed to bs_set_data()}
}
\value{
tag
}
\description{
Embed a tooltip into an html element
}
\examples{
library("htmltools")
tags$button("button") \%>\%
  bs_embed_tooltip(title = "I'm a tooltip")
}

