% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_modal.R
\name{bs_modal}
\alias{bs_modal}
\title{Make a bootstrap modal window}
\usage{
bs_modal(id, title, ..., size = c("large", "medium", "small"))
}
\arguments{
\item{id}{character, unique identifier for the modal window}

\item{title}{character (or HTML), to put in as the title of the modal window}

\item{...}{elements to include in the body}

\item{size}{character, how big to make the window}
}
\value{
\code{\link[htmltools]{div}} containing the code for the modal window
}
\description{
You need to include the \code{id} as the \code{`data-target`} of a button or link.
}

