% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_modal_helpify.R
\name{bs_modal_helpify}
\alias{bs_modal_helpify}
\title{Wraps an input with modal window}
\usage{
bs_modal_helpify(input, bs_modal)
}
\arguments{
\item{input, }{a Shiny input of some sort}

\item{bs_modal, }{a bs_modal window}
}
\description{
may work only for inputs with simple labels
}

