% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_panel.R
\name{bs_panel_collapse}
\alias{bs_panel_collapse}
\title{wrapper for collapseable bootstrap panel}
\usage{
bs_panel_collapse(..., id = NULL, title = NULL, is_open = TRUE,
  bstype = c("primary", "success", "info", "warning", "danger", "default"),
  use_bstype_open = FALSE)
}
\arguments{
\item{...}{UI elements to include in the panel}

\item{id}{character, id for the element}

\item{title}{character, title for the panel (maybe this can be html, not guaranteed)}

\item{is_open}{boolean, indicates if this div will be initially open (or closed)}

\item{bstype}{character, bootstrap type}

\item{use_bstype_open}{logical, indicates if to use the default tuype}
}
\value{
a panel div
}
\description{
wrapper for collapseable bootstrap panel
}

