% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Creates a collapsible div}
\usage{
collapse(id)
}
\arguments{
\item{id}{character, unique id corresponds to \code{id_collapse}}
}
\value{
\code{shiny.tag} div
}
\description{
Creates a collapsible div
}

