% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_append}
\alias{collapse_append}
\title{Appends a tag to a collapsible div}
\usage{
collapse_append(tag_collapse, tag_child)
}
\arguments{
\item{tag_collapse}{\code{\link[htmltools]{tag}}, made using \code{\link{collapse()}}}

\item{tag_child}{\code{\link[htmltools]{tag}}}
}
\value{
\code{\link[htmltools]{tag}}, modified copy of \code{tag_collapse}
}
\description{
This is syntactic sugar for htmltools \code{\link[htmltools]{tagAppendChild}}
}

