% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_control}
\alias{collapse_control}
\title{Modifies a tag to control a collpasible div}
\usage{
collapse_control(tag, id_target)
}
\arguments{
\item{tag}{\code{shiny.tag} either link or button}

\item{id_target}{character, id for the div to be controlled}
}
\value{
\code{shiny.tag} of the same type as \code{tag}
}
\description{
Works on either a button (\code{<button>}) or a link (\code{<a>}).
}
\details{
Adds attribute: \code{data-toggle="collapse"}.

If this is a button, also adds attribute \code{data-target="#[id_target]"}

If this is a link, also adds attributes \code{role="button"}, \code{href="#[id_target]"}.
}
\seealso{
\url{https://getbootstrap.com/javascript/#collapse}
}

