% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_sidebar_definition.R
\name{collapse_main_panel}
\alias{collapse_main_panel}
\title{Define a collapse_main_panel}
\usage{
collapse_main_panel(...)
}
\arguments{
\item{...}{UI elements to put into a `collapse_main_panel` body;
each top-level tag's will be inspected for a class of bootstrap panel,
and if there is an attribute to direct a bootstrap type for open panels.}
}
\value{
Function to create a div of main-panel elements. This function will be evaluated
  in \code{\link{collapse_sidebar_layout}}
}
\description{
Works in conjunction with \code{\link{bs_panel}} and \code{\link{bs_panel_collapse}}.
}

