% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_collapse.R
\name{collapse_panel}
\alias{collapse_panel}
\title{collapse_panel}
\usage{
collapse_panel(title = "", control = htmltools::tagList(),
  display = htmltools::tagList())
}
\arguments{
\item{title}{stuff (coercible to HTML) to appear in panel-heading}

\item{control}{shiny \code{tagList} of UI elements to appear in control panel}

\item{display}{shiny \code{tagList} of UI elements to appear in target panel}
}
\value{
list to be used by \code{collapse_panel_set}
}
\description{
creates an element to appear in a \code{collapse_panel_set}
}

