% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_collapse.R
\name{collapse_panel_set}
\alias{collapse_panel_set}
\title{collapse_panel_set}
\usage{
collapse_panel_set(..., widths = c(3, 9))
}
\arguments{
\item{...}{\code{collpase_panel} elements}

\item{widths}{numeric vector, length 2, indicating the widths of the columns for control and target}
}
\value{
shiny \code{fluidRow}
}
\description{
creates a fluidRow
}

