% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_sidebar_definition.R
\name{collapse_sidebar_layout}
\alias{collapse_sidebar_layout}
\title{Defines a function to composes panels into a collapse_sidebar_layout}
\usage{
collapse_sidebar_layout(collapse_sidebar_panel, collapse_main_panel, id_layout,
  bstype_closed = NULL)
}
\arguments{
\item{collapse_sidebar_panel}{function that returns `collapse_sidebar_panel` element}

\item{collapse_main_panel}{function that returns `main_collapse_panel` element}

\item{id_layout}{character, identifier for the sidebar_collapse_layout}

\item{bstype_closed}{character, bootstrap type for closed panels}
}
\value{
function that returns a \code{collapse_sidebar_layout} list
}
\description{
Defines a function to composes panels into a collapse_sidebar_layout
}

