% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_sidebar_definition.R
\name{collapse_sidebar_panel}
\alias{collapse_sidebar_panel}
\title{Define a collapse_sidebar_panel}
\usage{
collapse_sidebar_panel(..., title)
}
\arguments{
\item{...}{UI elements to put into a `collapse_sidebar_panel` body}

\item{title}{character or HTML, title}
}
\value{
Function to create a div of sidebaar-panel. This function will be evaluated
  in \code{\link{collapse_sidebar_layout}}
}
\description{
Works in conjunction with \code{\link{bs_panel}} and \code{\link{bs_panel_collapse}}.
}

