% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_sidebar_definition.R
\name{collapse_sidebar_set}
\alias{collapse_sidebar_set}
\title{Composes collapse_sidebar_layouts vertically into a collapse_sidebar_set}
\usage{
collapse_sidebar_set(id_set, ..., width = 3, position = c("left", "right"),
  bstype_open = c("success", "info", "warning", "danger", "default",
  "primary"), bstype_closed = c("primary", "success", "info", "warning",
  "danger", "default"))
}
\arguments{
\item{id_set}{character, identifier for the sidebar_collapse_set}

\item{...}{`collapse_sidebar_layout` functions}

\item{width}{integer, width (number of bootstrap columns) of the sidebar column}

\item{position}{character, position of sidebar column}

\item{bstype_open}{character, bootstrap type for open panels}

\item{bstype_closed}{character, bootstrap type for closed panels}
}
\value{
\code{\link[shiny]{fluidRow}} where columns are defined below, depending on `position`
}
\description{
Composes collapse_sidebar_layouts vertically into a collapse_sidebar_set
}

