% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_target}
\alias{collapse_target}
\title{Wraps a tag to make it collapsible}
\usage{
collapse_target(tag, id)
}
\arguments{
\item{tag}{\code{shiny.tag} to be collapsed}

\item{id}{character, unique id corresponds to \code{id_target}}
}
\value{
\code{shiny.tag}
}
\description{
Works on any tag - wraps it in a \code{<div>},
sets the attributes \code{class="collpase"} and \code{id="[id]"}
}

