% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_video.R
\name{embed_vimeo}
\alias{embed_vimeo}
\title{embed_vimeo}
\usage{
embed_vimeo(vimeo_id)
}
\arguments{
\item{vimeo_id}{character for vimeo_id}
}
\value{
html code describing div
}
\description{
creates a centered div for embedding a vimeo
}

