% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_video.R
\name{embed_youtube}
\alias{embed_youtube}
\title{creates a centered div for embedding a video from YouTube}
\usage{
embed_youtube(id, width = 500, height = 281)
}
\arguments{
\item{id}{character, id}

\item{width}{numeric, width of iframe (pixels)}

\item{height}{numeric, height of iframe (pixels)}
}
\value{
html code describing div
}
\description{
creates a centered div for embedding a video from YouTube
}
\examples{
embed_youtube("dQw4w9WgXcQ")

}

