% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R, R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs, rhs}{An object and a function to apply to it}
}
\description{
Like dplyr, bsplus also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.

Import mechanism taken from
https://github.com/hadley/modelr/blob/master/R/utils.R
}
\keyword{internal}

