% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bsopts.R
\name{set_bsdata}
\alias{set_bsaria}
\alias{set_bsdata}
\title{Sets Bootstrap data- and aria- attributes}
\usage{
set_bsdata(tag, ...)

set_bsaria(tag, ...)
}
\arguments{
\item{tag}{htmltools \code{\link[htmltools]{tag}}}

\item{...}{named arguments used to set the attributes of \code{tag}}
}
\value{
htmltools \code{\link[htmltools]{tag}}
}
\description{
Helper function to manage attributes for Bootstrap's Javascript components.
}
\details{
One of the mechanisms used by the API for Boostrap Javascript-components is
an html elements' attributes. These attribute names are prefixed with
\code{"data-"} or \code{"aria-"}, depending on the function.
When expressed in html, attributes themselves have the properties:

\itemize{
  \item Logical values are expressed as \code{"true"} or \code{"false"}.
  \item Time durations are expressed as number of milliseconds.
  \item Vector (non scalar) values are expressed in a space-delimited list.
}

The purpose of this function is to let you express these values in ways
familiar to you as an R user. For example:

\itemize{
  \item Logical values can be expressed as \code{TRUE} or \code{FALSE}.
  \item Time durations can be expressed using lubridate durations.
  \item Vector (non scalar) values can be expressed as vectors.
}
}
\examples{
library("htmltools")
library("lubridate")
tags$div() \%>\%
  set_bsdata(target = "#target", delay = dseconds(1)) \%>\%
  set_bsaria(expanded = FALSE)
}
\seealso{
\href{http://getbootstrap.com/javascript}{Boostrap Javascript Components}
}

