% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bsopts.R
\name{set_bsopts}
\alias{set_bsopts}
\title{Sets Bootstrap data-attributes}
\usage{
set_bsopts(tag, ...)
}
\arguments{
\item{tag}{htmltools \code{\link[htmltools]{tag}}}

\item{...}{named arguments used to set the attributes of \code{tag}}
}
\value{
htmltools \code{\link[htmltools]{tag}}
}
\description{
Helper function to manage attributes for Bootstrap's Javascript components.
}
\details{
One of the mechanisms used by the API for Boostrap Javascript-components is
an html elements' attributes. These attribute names are prefixed with
\code{"data-"}. The expressed in html, attributes themselves have the
properties:

\itemize{
  \item Logical values are expressed as \code{"true"} or \code{"false"}.
  \item Time durations are expressed as number of milliseconds.
  \item Vector (non scalar) values are expressed in a space-delimited list.
}

The purpose of this function is to let you express these values in ways
familiar to you as an R user. For example:

\itemize{
  \item Logical values can be expressed as \code{TRUE} or \code{FALSE}.
  \item Time durations can be expressed using lubridate durations.
  \item Vector (non scalar) values can be expressed as vectors.
}
}
\seealso{
\href{http://getbootstrap.com/javascript}{Boostrap Javascript Components}
}

