% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinyInput_label_embed}
\alias{shinyInput_label_embed}
\title{Embed a tag into the label of a Shiny input}
\usage{
shinyInput_label_embed(shinyInput, tag)
}
\arguments{
\item{shinyInput}{Shiny input, such as \code{shiny::\link[shiny]{numericInput}}}

\item{tag}{\code{htmltools::\link[htmltools]{tag}}, usually a link}
}
\value{
Shiny input, modified copy of \code{shinyInput}
}
\description{
Embed a tag into the label of a Shiny input
}
\examples{
library("shiny")

numericInput(inputId = "foo", label = "Enter a number", value = 0) \%>\%
  shinyInput_label_embed(
    shiny_iconlink() \%>\%
    bs_embed_popover(title = "Number", content = "Not a complex number")
  )

}
\seealso{
\code{\link{shiny_iconlink}}
}

