% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{attach_collapse}
\alias{attach_collapse}
\title{Attaches collpase control to a tag}
\usage{
attach_collapse(tag, id_collapse)
}
\arguments{
\item{tag}{\code{shiny.tag} either link or button}

\item{id_target}{character, id for the div to be controlled}
}
\value{
\code{shiny.tag} of the same type as \code{tag}
}
\description{
Works on either a button (\code{<button>}) or a link (\code{<a>}).
}
\seealso{
\url{https://getbootstrap.com/javascript/#collapse}
}

