% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{attach_collapse}
\alias{attach_collapse}
\title{Attaches a collapsible div to a control}
\usage{
attach_collapse(tag, id_collapse)
}
\arguments{
\item{tag}{htmltools \code{\link[htmltools]{tag}},}

\item{id_collapse}{character, id for the collapsible div}
}
\value{
\code{\link[htmltools]{tag}}, modified copy of \code{tag}
}
\description{
Typical to use either a button (\code{<button>}) or a link (\code{<a>}).
}
\details{
You may attach a collapseable div to more than one control.
}
\examples{
library("htmltools")
tags$button("She Loves You") \%>\%
  collapse_attach("id_yeah")

}
\seealso{
\url{https://getbootstrap.com/javascript/#collapse}

Other collapse.functions: \code{\link{collapse}}
}

