% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R, R/append.R, R/set_opts.R
\name{bs_accordion}
\alias{bs_accordion}
\alias{bs_append.bsplus_accordion}
\alias{bs_set_opts.bsplus_accordion}
\title{Builds an accordion panel-group.}
\usage{
bs_accordion(id)

\method{bs_append}{bsplus_accordion}(tag, title, content, ...)

\method{bs_set_opts}{bsplus_accordion}(tag, panel_type = "primary",
  use_heading_link = TRUE, ...)
}
\arguments{
\item{id}{character, id for div}

\item{tag}{\code{shiny.tag}, accordion element}

\item{title}{HTML or text, used for the panel heading}

\item{content}{HTML or \code{\link[htmltools]{tagList}}, used for the panel body.}

\item{...}{other args}

\item{panel_type}{character, one of the standard Bootstrap types
\code{c("default", "primary", "success", "info", "warning", "danger")}}

\item{use_heading_link}{logical, indicates whether to use the entire panel heading
as a link. If \code{FALSE}, uses construction shown in
\url{http://getbootstrap.com/javascript/#collapse-example-accordion};
if \code{TRUE}, makes the entire panel heading clickable.}
}
\value{
\code{bsplus_accordion} object
}
\description{
An accordion is a group of panels where, at most, one panel body is visible.
}
\details{
This constructor takes a single argument, an id.
It is your responsibility to ensure that id is unique
among html elements in your page. If you have non-unique id's, strange things may
happen as your page is rendered.

The constructor is a part of a family of function to set the options for
panels, and a function to append a panel to the panel group.

All of these functions return a \code{bsplus_accordion} object, so you can
compose an accordion by piping. The \code{bsplus_accordion} class inherits
from htmltools' \code{shiny.tag} class.

The \code{bs_set_opts} method is used to modify panels that have yet to
be appended. There are two options that can be set for accordions:

\describe{
  \item{\code{panel_type}}{One of \code{c("default", "primary", "success", "info", "warning", "danger")}.
  These are the standard Bootstrap types; the default is \code{"default"}.}
  \item{\code{use_heading_link}}{Indicates if the entire panel-heading shall be
  made clickable; the default is \code{FALSE}.}
}

The \code{append} method is used to build and append a panel to an accordion. Such a
panel is built using two additional arguments:

\describe{
  \item{\code{title}}{Usually text for the panel heading, but HTML can be used as well}
  \item{\code{content}}{Content for the panel body, can be HTML or an htmltools
  \code{\link[htmltools]{tagList}}}
}
}
\examples{
bs_accordion(id = "meet_the_beatles") \%>\%
  bs_set_opts(panel_type = "success", use_heading_link = TRUE) \%>\%
  bs_append(title = "John Lennon", content = "Rhythm guitar, vocals") \%>\%
  bs_set_opts(panel_type = "info") \%>\%
  bs_append(title = "Paul McCartney", content = "Bass guitar, vocals")

}
\seealso{
\url{http://getbootstrap.com/javascript/#collapse-example-accordion}
}

