% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_sidebar.R, R/append.R, R/set_opts.R
\name{bs_accordion_sidebar}
\alias{bs_accordion_sidebar}
\alias{bs_append.bsplus_accordion_sidebar}
\alias{bs_set_opts.bsplus_accordion_sidebar}
\alias{use_bs_accordion_sidebar}
\title{framework for accordion-sidebar}
\usage{
bs_accordion_sidebar(id, spec_side = c(width = 4, offset = 0),
  spec_main = c(width = 8, offset = 0), position = c("left", "right"))

use_bs_accordion_sidebar()

\method{bs_append}{bsplus_accordion_sidebar}(tag, title_side, content_side,
  content_main, ...)

\method{bs_set_opts}{bsplus_accordion_sidebar}(tag,
  panel_type_active = "success", panel_type_inactive = "primary", ...)
}
\arguments{
\item{id}{character, id for div and root it for included elements}

\item{spec_side}{numeric, column specification for sidebar panels}

\item{spec_main}{numeric, column specification for main panels}

\item{position}{character, indicates where to put the sidebar panels with
repspect to the main panels}

\item{tag}{\code{shiny.tag}, accordion-sidebar element}

\item{title_side}{character or HTML, stuff to put into sidebar title}

\item{content_side}{character or HTML, stuff to put into sidebar content}

\item{content_main}{character or HTML, stuff to put into main content}

\item{...}{other args}

\item{panel_type_active}{character, indicated bootstrap type for active-panel header}

\item{panel_type_inactive}{character, indicated bootstrap type for inactive-panel header}
}
\value{
\code{bsplus_accsidebar} object
}
\description{
framework for accordion-sidebar
}
\examples{
bs_accordion_sidebar(id = "meet_the_beatles") \%>\%
  bs_append(
    title_side = "John Lennon",
    content_side = "Rhythm guitar, vocals",
    content_main = "Dear Prudence"
  ) \%>\%
  bs_append(
    title_side = "Paul McCartney",
    content_side = "Bass guitar, vocals",
    content_main = "Blackbird"
  )
\dontrun{
use_bs_accordion_sidebar()
}
}

