% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{bs_button}
\alias{bs_button}
\title{Button}
\usage{
bs_button(label, button_type = c("default", "primary", "success", "info",
  "warning", "danger"), ...)
}
\arguments{
\item{label}{character (HTML), button label}

\item{button_type}{character, one of the standard Bootstrap types}

\item{...}{attributes (named arguments) and children (unnamed arguments)
of the button, passed to \code{htmltools::\link[htmltools]{tag}}}
}
\value{
\code{htmltools::\link[htmltools]{tag}}, \code{<button/>}
}
\description{
This function makes it a little easier to make Bootstrap-friendly buttons;
it wraps the [htmltools::tags()] function for buttons.
}
\examples{
bs_button("Click me", button_type = "primary")
}
\seealso{
\url{http://getbootstrap.com/css/#buttons}
}
