% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{bs_collapse}
\alias{bs_collapse}
\title{Creates a collapsible div}
\usage{
bs_collapse(id, content = NULL)
}
\arguments{
\item{id}{character, id for div}

\item{content}{htmltools \code{\link[htmltools]{tagList}} or html, content to be
put into the div}
}
\value{
\code{\link[htmltools]{tag}} div
}
\description{
When a page is renedered, this div will be hidden initialy, but can be revealed
using a control to which you will have attached this div's id.
}
\details{
It is your responsibility to ensure that id is unique
among html elements in your page. If you have non-unique id's, strange things may
happen as your page is rendered.

This div's id may be attached to more than one control, see
\code{\link{attach_collapse}}.
}
\examples{
library("htmltools")
bs_collapse(id = "id_yeah", tags$p("Yeah Yeah Yeah"))
tags$button("She Loves You") \%>\%
  bs_attach_collapse("id_yeah")

}
\seealso{
\url{https://getbootstrap.com/javascript/#collapse}

Other collapse.functions: \code{\link{bs_attach_collapse}}
}

