% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{bs_collapse}
\alias{bs_attach_collapse}
\alias{bs_collapse}
\title{Collapsible element}
\usage{
bs_collapse(id, content = NULL, show = FALSE)

bs_attach_collapse(tag, id_collapse)
}
\arguments{
\item{id}{character, unique id for collapsible div}

\item{content}{character (HTML) or \code{htmltools::\link[htmltools]{tagList}},
content to be put into the div}

\item{show}{logical, indicates if element is shown initially}

\item{tag}{\code{htmltools::\link[htmltools]{tag}},
button or link to which to attach collapsible div}

\item{id_collapse}{character, id of collapsible div to attach}
}
\value{
\describe{
  \item{\code{bs_collapse()}}{\code{htmltools::\link[htmltools]{tag}}, div}
  \item{\code{bs_attach_collapse()}}{\code{htmltools::\link[htmltools]{tag}},
    modified copy of \code{tag}}
}
}
\description{
This is useful for content that you may wish to be hidden when the page is
initialized, but that can be revealed (and subsequently hidden)
by clicking a button or a link.
}
\details{
There are two parts to this system:

\enumerate{
  \item{A collapsible div, created using \code{bs_collapse()}}
  \item{At least one button or link to which the id of the
  collabsible div is attached, using \code{bs_attach_collapse()}}
}

The verb "attach" is used to signify that we are attaching the id of our collapsible
div to the tag in question (likely a button or a link). This implies that
you can attach (the id of) a collapsible div to more than one button or link.

It is your responsibility to ensure that id is unique
among html elements in your page. If you have non-unique id's, strange things may
happen to your page.
}
\examples{
library("htmltools")

bs_collapse(id = "id_yeah", tags$p("Yeah Yeah Yeah"))

tags$button("She Loves You") \%>\%
  bs_attach_collapse("id_yeah")

}
\seealso{
\url{https://getbootstrap.com/javascript/#collapse}
}

