% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{bs_modal}
\alias{bs_attach_modal}
\alias{bs_modal}
\alias{bs_modal_closebutton}
\title{Create and attach a modal window}
\usage{
bs_modal(id, title, body, footer = bs_modal_closebutton(title = "Close"),
  size = c("medium", "large", "small"))

bs_modal_closebutton(title = "Close")

bs_attach_modal(tag, id_modal)
}
\arguments{
\item{id}{character, unique id for modal window to create}

\item{title}{character, title for the modal or close-button}

\item{body}{character, content of the body, can be HTML fragment}

\item{footer}{character, HTML fragment for footer}

\item{size}{character, size of the modal}

\item{tag}{character, HTML element to which to attach modal}

\item{id_modal}{character, unique id of modal window to attach}
}
\value{
HTML fragment
}
\description{
You can attach an modal window to more than one element.
}
\examples{
library("htmltools")

bs_modal(id = "modal", title = "I'm a modal", body = "Yes, I am.")
tags$button(type = "button", class = "btn btn-default", "Click for modal") \%>\%
  bs_attach_modal(id_modal = "modal")

}

