% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_panel.R
\name{bs_panel}
\alias{bs_panel}
\title{wrapper for bootstrap panel}
\usage{
bs_panel(type = "primary", title = NULL, ...)
}
\arguments{
\item{type, }{character bootstrap class,
one of: c("default", "primary", "success", "info", "warning", "danger")}

\item{title, }{cracter, title for the panel (maybe this can be html, not guaranteed)}

\item{...}{UI elements to include in the panel}
}
\description{
wrapper for bootstrap panel
}

