% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_opts.R
\name{bs_set_opts}
\alias{bs_set_opts}
\alias{bs_set_opts.default}
\alias{bs_set_opts.shiny.tag}
\title{Set the options}
\usage{
bs_set_opts(tag, ...)

\method{bs_set_opts}{default}(tag, ...)

\method{bs_set_opts}{shiny.tag}(tag, ...)
}
\arguments{
\item{tag}{\code{htmltools:shinytag}}

\item{...}{additional arguments}
}
\value{
modified copy of \code{tag}
}
\description{
This would be used when we want set options for adding items
}

