% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Creates a collapsible div}
\usage{
collapse(id, content = NULL)
}
\arguments{
\item{id}{character, unique id for collapsible div}

\item{content}{htmltools \code{\link[htmltools]{tagList}}}
}
\value{
\code{\link[htmltools]{tag}} div
}
\description{
It is your responsibility to ensure that id is unique
among html elements in your page. If you have non-unique id's, strange things may
happen as your page is rendered.
}
\examples{
library("htmltools")
collapse(id = "id_yeah", tags$p("Yeah Yeah Yeah"))

}
\seealso{
\url{https://getbootstrap.com/javascript/#collapse}

Other collapse.functions: \code{\link{attach_collapse}}
}

