% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Creates a collapsible div}
\usage{
collapse(id)
}
\arguments{
\item{id}{character, unique id for collapsible div}
}
\value{
\code{\link[htmltools]{tag}} div
}
\description{
Important to note that it is your responsibility to ensure that id is unique
among html elements in your page. If you have non-unique id's, strange things may
happen as your page is rendered.
}
\examples{
collapse(id = "id_yeah")

}
\seealso{
\url{https://getbootstrap.com/javascript/#collapse}

Other collapse.functions: \code{\link{attach_collapse}},
  \code{\link{collapse_append}}
}

