% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_append}
\alias{collapse_append}
\title{Appends a tag to a collapsible div}
\usage{
collapse_append(tag_collapse, tag_child)
}
\arguments{
\item{tag_collapse}{\code{\link[htmltools]{tag}}, made using \code{\link{collapse()}}}

\item{tag_child}{\code{\link[htmltools]{tag}} to be appended to \code{tag_collapse}}
}
\value{
\code{\link[htmltools]{tag}}, modified copy of \code{tag_collapse}
}
\description{
This is a thin wrapper to htmltools \code{\link[htmltools]{tagAppendChild}}
}
\examples{
collapse(id = "id_yeah") \%>\%
  collapse_append("Yeah Yeah Yeah")

}
\seealso{
\url{https://getbootstrap.com/javascript/#collapse}

Other collapse.functions: \code{\link{attach_collapse}},
  \code{\link{collapse}}
}

