% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_html_fragment.R
\name{render_html_fragment}
\alias{render_html_fragment}
\title{render an html fragment}
\usage{
render_html_fragment(input, output_format = rmarkdown::html_fragment(), ...)
}
\arguments{
\item{input}{character, path to input file}

\item{output_format}{rmarkdown format, provided so you can specify arguments}

\item{...}{other arguments passed to \code{rmarkdown::\link{render}}}
}
\value{
HTML fragment (unless another format specified)
}
\description{
This is a wrapper around the \code{rmarkdown::\link{render}} function.
The principal difference is that the function returns an HTML fragment so
that it can be used, for example, in a modal window or a popover.
}
\examples{
my_file <- system.file("markdown", "modal.md", package = "bsplus")
render_html_fragment(my_file)
}

